Moves the DefaultTenantResolver to a separate file
